Refactor to prep for testing
